refactor: merging conflicts, bring all branches back to dev
