+++ src/routes/Signup.jsx
@@ -1,62 +1,125 @@
+import { useRef } from "react";
+import { firestore } from "../firebase";
+import { addDoc, collection } from "firebase/firestore";
+import { getAuth, createUserWithEmailAndPassword } from "firebase/auth";
+
 function Signup() {
+  const auth = getAuth();
+  const firstNameRef = useRef();
+  const lastNameRef = useRef();
+  const emailRef = useRef();
+  const passwordRef = useRef();
+  const confirmPasswordRef = useRef();
+  const userNameRef = useRef();
+
+  const ref = collection(firestore, "users");
+  const handleSave = async (e) => {
+    e.preventDefault();
+    let data = {
+      firstName: firstNameRef.current.value,
+      lastName: lastNameRef.current.value,
+      email: emailRef.current.value,
+      password: passwordRef.current.value,
+      confirmPassword: confirmPasswordRef.current.value,
+      userName: userNameRef.current.value,
+    };
+    try {
+      await createUserWithEmailAndPassword(auth, data.email, data.password);
+      addDoc(ref, data);
+      console.log("User created");
+    } catch (err) {
+      console.log(err);
+    }
+  };
+
   return (
     <div>
-      <div class="container mt-5">
+      <div className="container mt-5">
         <h2>Registration Form</h2>
-        <form>
-          <div class="mb-3">
-            <label for="firstName" class="form-label">
+        <form onSubmit={handleSave}>
+          <div className="mb-3">
+            <label htmlFor="firstName" className="form-label">
               First Name
             </label>
             <input
               type="text"
-              class="form-control"
+              className="form-control"
               id="firstName"
               name="firstName"
               required
+              ref={firstNameRef}
             />
           </div>
-          <div class="mb-3">
-            <label for="lastName" class="form-label">
+          <div className="mb-3">
+            <label htmlFor="lastName" className="form-label">
               Last Name
             </label>
             <input
               type="text"
-              class="form-control"
+              className="form-control"
               id="lastName"
               name="lastName"
               required
+              ref={lastNameRef}
             />
           </div>
-          <div class="mb-3">
-            <label for="username" class="form-label">
+          <div className="mb-3">
+            <label htmlFor="email" className="form-label">
+              Email
+            </label>
+            <input
+              type="email"
+              className="form-control"
+              id="email"
+              name="email"
+              required
+              ref={emailRef}
+            />
+          </div>
+          <div className="mb-3">
+            <label htmlFor="username" className="form-label">
               Username
             </label>
             <input
               type="text"
-              class="form-control"
+              className="form-control"
               id="username"
               name="username"
               required
+              ref={userNameRef}
             />
           </div>
-          <div class="mb-3">
-            <label for="password" class="form-label">
+          <div className="mb-3">
+            <label htmlFor="password" className="form-label">
               Password
             </label>
             <input
               type="password"
-              class="form-control"
+              className="form-control"
               id="password"
               name="password"
               required
+              ref={passwordRef}
             />
           </div>
-          <button type="submit" class="btn btn-primary">
+          <div className="mb-3">
+            <label htmlFor="confirmPassword" className="form-label">
+              confirmPassword
+            </label>
+            <input
+              type="confirmPassword"
+              className="form-control"
+              id="confirmPassword"
+              name="confirmPassword"
+              required
+              ref={confirmPasswordRef}
+            />
+          </div>
+          <button type="submit" className="btn btn-primary">
             Register
           </button>
         </form>
-        <p class="mt-3">
+        <p className="mt-3">
           Already have an account? <a href="#">Login</a>
         </p>
       </div>
